requirements.txt
# Backend Python Dependencies
firebase-admin>=6.0.0
python-dotenv>=1.0.0

# Face Recognition Dependencies (Recog_Face)
tensorflow==2.13.0
numpy==1.24.3
opencv-python>=4.8.0
pillow>=10.0.0
supabase>=2.0.0
ultralytics>=8.0.0
omegaconf>=2.3.0
flask>=3.0.0
flask-cors>=4.0.0
